+++ bookvoice/io/__init__.py
@@ -5,11 +5,13 @@ interfaces used by the pipeline.
 """
 
 from .chapter_splitter import ChapterSplitter
+from .epub_text_extractor import EpubTextExtractor
 from .pdf_outline_extractor import PdfOutlineChapterExtractor
 from .pdf_text_extractor import PdfTextExtractor
 from .storage import ArtifactStore
 
 __all__ = [
+    "EpubTextExtractor",
     "PdfTextExtractor",
     "PdfOutlineChapterExtractor",
     "ChapterSplitter",